W-13871014: Improve test coverage (#159)
* - Update POM
- remove unnecessary methods
- add additional MUnit tests

* - add additional tests for Xslt
- remove unused contructor
- Fix a bug when transforming Integers for Xslt context

* add empty constructor back

+++ src/main/java/org/mule/module/xml/internal/operation/XsltOperation.java
@@ -113,7 +113,7 @@ public class XsltOperation extends PooledTransformerOperation<String, XsltTransf
       } else if (value instanceof InputStream) {
         xdmValue = new XdmAtomicValue(IOUtils.toString((InputStream) value));
       } else if (value instanceof Integer) {
-        xdmValue = new XdmAtomicValue((Long) value);
+        xdmValue = new XdmAtomicValue((Integer) value);
       } else if (value instanceof Boolean) {
         xdmValue = new XdmAtomicValue((Boolean) value);
       } else if (value instanceof Long) {
